refactor(core): group watch subsystem
