+++ src/core/loaders.ts
@@ -21,7 +21,7 @@ import {
 } from "./vcs";
 import type { VcsCatalog } from "./vcs/types";
 import { buildFilesystemUntrackedDiffFile } from "./vcs/untracked";
-import { computeWatchSignature } from "./watch";
+import { computeWatchSignature } from "./watch/signature";
 import type {
   AppBootstrap,
   AgentContext,
