+++ src/core/watch/signature.ts
@@ -1,8 +1,8 @@
 import fs from "node:fs";
 import { resolve } from "node:path";
-import { createVcsWatchSignature, getConfiguredVcsAdapter, operationFromInput } from "./vcs";
-import type { CliInput } from "./types";
-import type { VcsCatalog } from "./vcs/types";
+import { createVcsWatchSignature, getConfiguredVcsAdapter, operationFromInput } from "../vcs";
+import type { CliInput } from "../types";
+import type { VcsCatalog } from "../vcs/types";
 
 /** Format one file stat into a stable signature fragment, or mark the path missing. */
 function statSignature(path: string) {
